Exit on error
As reported in the upstream project[1], if the main Unison process
exists, like when the SSH connection is dropped, fsmonitor keeps
printing errors in an infinite loop. This, in turn, uses tons of CPU and
memory, to the point of exhausting system resources. Exiting whenever it
detects an error prevents this issue.

[1]: https://github.com/autozimu/unison-fsmonitor/issues/27